Chore: 구조 변화 (#11)
* feat: 스타일 조정

* fix: 동적 랜더링

* fix: 오타 제거

* chore: jwt에서 다른거 시도

* chore: 구조 변화

변화
